Improved the script for removing the nested node_modules directory.
