+++ package.json
@@ -148,7 +148,7 @@
     "manual": "node --max_old_space_size=8192 node_modules/@ckeditor/ckeditor5-dev-tests/bin/test-manual.js",
     "manual:verify": "node ./scripts/web-crawler/index.js --manual",
     "bootstrap": "yarn install",
-    "clean": "node ./scripts/remove-lock-files.js && rm -rf ./node_modules && rm -rf ./external/*/node_modules",
+    "clean": "node ./scripts/revert-yarn-install.js",
     "reset": "rm -rf ./node_modules && yarn run bootstrap",
     "reinstall": "yarn run clean && yarn run bootstrap",
     "docs": "node --max-old-space-size=4096 ./scripts/docs/build-docs.js",
